refactor FakeDataStore to provide integration checking
