--- core_chat_data/src/commonTest/kotlin/net/thechance/mena/core_chat/contacts/fakes/FakeDataStore.kt
@@ -1,22 +1,0 @@
-package net.thechance.mena.core_chat.contacts.fakes
-
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.emptyPreferences
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-
-class FakeDataStore(
-    initial: Preferences = emptyPreferences()
-) : DataStore<Preferences> {
-
-    private val state = MutableStateFlow(initial)
-
-    override val data: Flow<Preferences> = state
-
-    override suspend fun updateData(transform: suspend (t: Preferences) -> Preferences): Preferences {
-        val newValue = transform(state.value)
-        state.value = newValue
-        return newValue
-    }
-}
